refactor : 메서드의 역할에 맞춰 코드를 결과 관련 메서드로 이관
