+++ src/main/java/lotto/LottoApplication.java
@@ -4,7 +4,7 @@ package lotto;
 import lotto.controller.LottoGenerator;
 import lotto.wrapper.Count;
 import lotto.wrapper.Money;
-import lotto.controller.MatchResult;
+import lotto.controller.CheckResult;
 import lotto.domain.Buyer;
 import lotto.domain.Lotto;
 import lotto.domain.LottoTicket;
@@ -18,7 +18,7 @@ public class LottoApplication {
     private static InputView input;
     private static Buyer buyer;
     private static List<LottoTicket> tickets;
-    private static MatchResult result;
+    private static CheckResult checker;
     static Count lottos;
     static Money amount;
     static Count lottosManual;
@@ -31,7 +31,7 @@ public class LottoApplication {
             init();
             buyLotto();
             confirmLotto();
-            result();
+            resultLotto();
         } catch (IllegalArgumentException iae) {
             iae.printStackTrace();
         } catch (Exception e) {
@@ -44,6 +44,7 @@ public class LottoApplication {
         input = new InputView();
         buyer = new Buyer();
     }
+
     private static void buyLottoManual() {
         // 수동 구매 수량을 결정.
         OutputView.printManual(); // 수동으로 구매 할 수를 입력하세요.
@@ -53,11 +54,13 @@ public class LottoApplication {
 
         OutputView.printSelectManual(); // 로또 수동 입력 하라는 프린트.
         lottosManualRaw = input.selectLottosManually(lottosManual);
-        tickets = LottoGenerator.rawToTicktes(lottosManualRaw); // 수동으로 입력하기.
+        tickets = LottoGenerator.rawToTickets(lottosManualRaw); // 수동으로 입력하기.
     }
+
     private static void buyLorroRandom() {
         tickets.addAll(buyer.buyLotto(randomAmount));// 자동 구매 수량에 따라 티켓 생성.
     }
+
     private static void buyLotto() {
         // 로또 입력
         // 금액 입력
@@ -68,16 +71,15 @@ public class LottoApplication {
         buyLottoManual();
         buyLorroRandom();
 
-        // 로또 실행
         // 구매
         // 구매 수량 확인
         OutputView.printBuyerTickets(lottosManual, lottosRandom); // 수동으로 a 개, 자동으로 b 개 구매했습니다.
-        result = new MatchResult(Count.sum(lottosManual, lottosRandom));
-        // 구매자 로또 정보 메시지
         OutputView.printBuyTicketsNumbers(tickets);
     }
 
     private static void confirmLotto() {
+        // 로또 체크 할 인스턴스 생성
+        checker = new CheckResult(Count.sum(lottosManual, lottosRandom));
         // 로또 당첨 번호
         OutputView.printLottoMsg();
         OutputView.printLottoNumbers(lotto.getLottoTicket());
@@ -86,13 +88,13 @@ public class LottoApplication {
         OutputView.printBonusNumber(lotto.getLottoTicket());
 
         // Match Lotto
-        result.countUpMatch(lotto, tickets);
+        checker.countUpMatch(lotto, tickets);
     }
 
-    private static void result() {
+    private static void resultLotto() {
         // return result to output view
-        List<Count> res = result.getResult();
-        Double resRevenue = result.calculateWinningRevenue();
+        List<Count> res = checker.getResult();
+        Double resRevenue = checker.calculateWinningRevenue();
         // 로또 결과 메시지
         OutputView.printResult(res);
         // 통계 출력 메시지
