+++ src/main/java/lotto/controller/CheckResult.java
