+++ src/test/java/lotto/controller/CheckResultTest.java
